Minor refactoring changes
Moved authorization logic to lib
Fixed prettier import order configuration
