+++ package.json
@@ -9,6 +9,7 @@
     "start": "next start",
     "lint": "next lint",
     "test": "jest --ci",
+    "prettify": "prettier --write .",
     "supastart": "supabase start",
     "supastop": "supabase stop"
   },
