+++ prettier.config.mjs
@@ -1,27 +1,38 @@
-/** @type {import('prettier').Config} */
-const prettierConfig = {
-  endOfLine: "lf",
+/** @type {import("@ianvs/prettier-plugin-sort-imports").PluginConfig} */
+const importConfig = {
   // prettier-ignore
   importOrder: [
-    "", "react", "", "<BUILTIN_MODULES>", "", "<THIRD_PARTY_MODULES>", "",
-    "^@/lib/(.*)$", "", "^@/components/(.*)$", "", "^@/ui/(.*)$", "",
-    "^(?!.*[.]css$)[./].*$", "", ".css$",
+    "", "^react$",
+    "", "<BUILTIN_MODULES>",
+    "", "<THIRD_PARTY_MODULES>",
+    "", "^@/(?!components|ui).*$",
+    "", "^@/components/.*$",
+    "", "^@/ui/.*$",
+    "", "^(?!.*[.]css$)[./].*$",
+    "", ".css$",
   ],
   importOrderTypeScriptVersion: "5.0.0",
-  jsonRecursiveSort: true,
-  jsxSingleQuote: false,
-  plugins: [
-    "prettier-plugin-sql",
-    "prettier-plugin-prisma",
-    "prettier-plugin-sort-json",
-    "@ianvs/prettier-plugin-sort-imports",
-    "prettier-plugin-tailwindcss",
-  ],
-  semi: false,
-  singleQuote: false,
-  tabWidth: 2,
+}
+
+/** @type {import("prettier-plugin-tailwindcss").PluginOptions} */
+const tailwindConfig = {
   tailwindFunctions: ["cva", "clsx", "cn"],
   tailwindStylesheet: "./src/app/globals.css",
 }
 
-export default prettierConfig
+/** @type {import("prettier").Config} */
+const config = {
+  ...importConfig,
+  ...tailwindConfig,
+
+  semi: false,
+  jsonRecursiveSort: true,
+
+  plugins: [
+    "prettier-plugin-sort-json",
+    "@ianvs/prettier-plugin-sort-imports",
+    "prettier-plugin-tailwindcss",
+  ],
+}
+
+export default config
