+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next"
 
 import "./globals.css"
 
-import Navbar from "@/components/Navbar"
+import Navbar from "@/components/navbar"
 
 export const metadata: Metadata = {
   title: "Car Lander",
