+++ src/lib/auth.config.ts
