+++ src/middleware.ts
@@ -1,5 +1,6 @@
-import authConfig from "@/auth.config"
 import NextAuth from "next-auth"
+
+import authConfig from "@/lib/auth.config"
 
 export default NextAuth(authConfig).auth
