+++ src/ui/avatar.tsx
@@ -1,7 +1,8 @@
 "use client"
 
-import * as AvatarPrimitive from "@radix-ui/react-avatar"
 import * as React from "react"
+
+import * as AvatarPrimitive from "@radix-ui/react-avatar"
 
 import { cn } from "@/lib/utils"
