+++ src/ui/button.tsx
@@ -1,6 +1,6 @@
-import { Slot } from "@radix-ui/react-slot"
 import * as React from "react"
 
+import { Slot } from "@radix-ui/react-slot"
 import { cva, type VariantProps } from "class-variance-authority"
 
 import { cn } from "@/lib/utils"
