+++ src/ui/label.tsx
@@ -1,8 +1,8 @@
 "use client"
 
-import * as LabelPrimitive from "@radix-ui/react-label"
 import * as React from "react"
 
+import * as LabelPrimitive from "@radix-ui/react-label"
 import { cva, type VariantProps } from "class-variance-authority"
 
 import { cn } from "@/lib/utils"
