+++ src/ui/separator.tsx
@@ -1,7 +1,8 @@
 "use client"
 
-import * as SeparatorPrimitive from "@radix-ui/react-separator"
 import * as React from "react"
+
+import * as SeparatorPrimitive from "@radix-ui/react-separator"
 
 import { cn } from "@/lib/utils"
