+++ tailwind.config.ts
@@ -4,11 +4,7 @@ import type { Config } from "tailwindcss"
 
 export default {
   darkMode: ["class"],
-  content: [
-    "./src/ui/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
-  ],
+  content: ["./src/{ui,components,app}/**/*.{js,ts,jsx,tsx,mdx}"],
   theme: {
     extend: {
       borderRadius: {
